requirements.txt
aldjemy==0.7.0
asgi-redis==1.4.1
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
backports.shutil-get-terminal-size==1.0.0
certifi==2018.4.16
channels==1.1.3
chardet==3.0.4
click==6.7
constantly==15.1.0
Cython==0.25.2
daphne==1.2.0
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.1.0
django-extensions==1.7.9
django-filter==1.0.4
django-toolbelt==0.0.1
djangorestframework==3.6.3
ecdsa==0.13
enum34==1.1.6
gunicorn==19.7.1
haikunator==2.1.0
httplib2==0.10.3
hyperlink==17.1.1
idna==2.5
incremental==17.5.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Markdown==2.6.8
msgpack-python==0.4.8
mysqlclient==1.3.0
oauthlib==2.0.2
olefile==0.44
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.3.1
Pygments==2.2.0
PyJWT==1.5.0
python-dotenv==0.8.2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
rsa==3.4.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
social-auth-core==1.4.0
SQLAlchemy==1.1.10
static3==0.7.0
traitlets==4.3.2
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
uWSGI==2.0.15
wcwidth==0.1.7
whitenoise==3.3.0
zope.interface==4.4.2
